* salt-review:
  fix: framework ordering of right side tree
  fix: identifier lsItems allow type and identifier value (not generated name)
  tests: additional unit tests
  tests: add voter unit tests
  build: do not do a build for just doc changes
  tests: do additional checks (lint, etc.) during build/test
  refactor: reliability improvements, code quality
  feat: additional logging and tests
  feat: add limit to number of rows fetched (50k)
  feat: check that JSON is valid and throw errors if not
  tests: revert to using standalone-chrome
  tests: update chome version
  tests: fix failed test
  tests: update for phpunit
  deps: update dependencies
  feat: add transactions to db changes
  refactor: use transactional migrations now that upstream issue is fixed
  deploy: use non-root user when running, add scanner when building

+++ core/src/Domain/FrontMatter/FrontMatterService.php
@@ -19,6 +19,9 @@ readonly class FrontMatterService
     public function deleteFrontMatter(FrontMatter $template): void
     {
         $delete = $this->em->find(FrontMatter::class, $template->getId());
+        if (null === $delete) {
+            return;
+        }
         $this->em->remove($delete);
         $this->em->flush();
     }
